requirements.txt
opencv-python
numpy
tqdm
faiss-gpu